setSubtotal was added in ItemData in order to get the subtotal automatically, dataType for ivaValue was changed, and RouterRestItem was added.

+++ src/main/java/co/com/onwheels/onwheelsb/infrastructure/drivenadapters/data/ProductData.java
@@ -50,7 +50,7 @@ public class ProductData {
     private Double unitaryPrice;
 
     @NotNull(message = "UnitaryPrice can't be null")
-    private Integer ivaValue;
+    private Double ivaValue;
 
     @NotNull(message = "Inventory can't be null")
     private Integer inventory;
